重构：重新调整程序结构
