+++ spider/__init__.py
@@ -1,0 +1,2 @@
+from spider.spider import get_search_results,get_custom_info
+from spider.spider_se import get_search_result
